% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{agg_climate_data}
\alias{agg_climate_data}
\alias{agg_climate_data_multiyear}
\title{Aggregate Climate Data}
\usage{
agg_climate_data(
  year,
  data_raster,
  climate_var,
  daily_agg,
  trans = "polynomial",
  trans_specs,
  geoweights_table,
  second_weights = FALSE
)

agg_climate_data_multiyear(
  years,
  data_raster,
  climate_var,
  daily_agg,
  trans = "polynomial",
  trans_specs,
  geoweights_table
)
}
\arguments{
\item{year}{The year to aggregate}

\item{data_raster}{The raster containing climate data}

\item{climate_var}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{trans}{A statistical transformation ('polynomial' currently supported)}

\item{trans_specs}{Specifications pertaining to selected transformation. For
'polynomial', this is an integer indicating the degree.}

\item{geoweights_table}{A table of weights which can be generated using
the function calc_geoweights()}

\item{second_weights}{A boolean indicating whether secondary weights should be used}

\item{years}{A vector containing the years to aggregate over}
}
\description{
Aggregate Climate Data
}
\section{Functions}{
\itemize{
\item \code{agg_climate_data_multiyear}: aggregate climate data over multiple years in parallel
}}

\examples{
agg_climate_data(2011, demo_prcp, "prcp", "sum", "polynomial", 3, demo_output_geoweights, second_weights = TRUE)
}
