% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{as.data.table.raster}
\alias{as.data.table.raster}
\title{Aggregate Climate Data}
\usage{
as.data.table.raster(
  x,
  row.names = NULL,
  optional = FALSE,
  xy = FALSE,
  inmem = raster::canProcessInMemory(x, 2),
  ...
)
}
\arguments{
\item{year}{The year to aggregate}

\item{data_raster}{The raster containing climate data}

\item{climate_var}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{trans}{A statistical transformation ('polynomial' currently supported)}

\item{trans_specs}{Specifications pertaining to selected transformation. For
'polynomial', this is an integer indicating the degree.}

\item{geoweights_table}{A table of weights which can be generated using
the function calc_geoweights()}

\item{second_weights}{A boolean indicating whether secondary weights should be used}
}
\description{
Aggregate Climate Data
}
\examples{
staggregate_polynomial(demo_prcp, "prcp", daily_agg = "sum", degree = 3, geoweights_table = demo_output_geoweights, second_weights = TRUE)
staggregate_spline(demo_prcp, "prcp", daily_agg = "sum", knot_locs = c(1,2,3), geoweights_table = demo_output_geoweights, second_weights = TRUE)
staggregate_bin(demo_prcp, "prcp", daily_agg = "sum", num_bins = 30, min = 5, geoweights_table = demo_output_geoweights, second_weights = TRUE)


}
