% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_overlay.R
\name{calc_geoweights}
\alias{calc_geoweights}
\title{Function to find spatial overlap between a raster and a set of polygons}
\usage{
calc_geoweights(data_source, input_polygons, polygon_id, weights_table = NULL)
}
\arguments{
\item{data_source}{the source of climate data (default is era5)}

\item{input_polygons}{a simple features polygon or multipolygon object}

\item{polygon_id}{the name of a column in the sf object representing a unique
identifier for each polygon}

\item{weights_table}{an optional data table of secondary weights}
}
\value{
a data.table of geoweights (area weighted raster/polygon overlap)
}
\description{
Function to find spatial overlap between a raster and a set of polygons
}
\examples{
calc_geoweights(data_era5, data_polygon, "countyfp", output_weights)
calc_geoweights(data_era5, data_polygon, "countyfp")

}
