% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_weights.R
\name{calc_raster_weights}
\alias{calc_raster_weights}
\title{Function that resamples a raster of interest to the ERA grid}
\usage{
calc_raster_weights(weights_raster, data_source, extent = "full")
}
\arguments{
\item{weights_raster}{a raster of a continuous variable, for example cropland coverage or population}

\item{data_source}{the source of climate data (default is era5)}

\item{extent}{an optional extent to crop the weights_raster to for faster processing}
}
\value{
a data.table of geoweights (area weighted raster/polygon overlap)
}
\description{
Function that resamples a raster of interest to the ERA grid
}
\examples{
calc_raster_weights(data_weights, data_era5)

}
