% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cropland_kansas_2011.R
\docType{data}
\name{cropland_kansas_2011}
\alias{cropland_kansas_2011}
\title{Cropland data for a portion of Kansas.}
\format{
An object of class \code{RasterLayer} of dimension 2000 x 2000 x 1.
}
\usage{
cropland_kansas_2011
}
\description{
A high resolution raster layer depicting cropland over an area in Kansas 0.5
degrees of latitude by 0.5 degrees of longitude.
}
\keyword{datasets}
