% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cropland_kansas_2015.R
\docType{data}
\name{cropland_kansas_2015}
\alias{cropland_kansas_2015}
\title{Cropland data for the state of Kansas.}
\format{
An object of class \code{RasterLayer} of dimension 12041 x 29853 x 1.
}
\usage{
cropland_kansas_2015
}
\description{
A high resolution raster layer depicting cropland over the state of Kansas
0.5 degrees of latitude by 0.5 degrees of longitude.
}
\keyword{datasets}
