% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cropland_nj_2015.R
\docType{data}
\name{cropland_nj_2015}
\alias{cropland_nj_2015}
\title{Cropland data for the state of New Jersey.}
\format{
An object of class \code{RasterLayer} of dimension 150 x 140 x 1.
}
\usage{
cropland_nj_2015
}
\description{
A raster depicting cropland in and around the state of New Jersey. The
original data is resolution 0.00025 latitude by 0.00025 longitude and comes
from the Potapov et. al. dataset available at
https://glad.umd.edu/dataset/croplands. This raster has been resampled to a
resolution of 0.025 latitude by 0.025 longitude to meet memory constraints.
}
\keyword{datasets}
