% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cropland_world_2003_era5.R
\docType{data}
\name{cropland_world_2003_era5}
\alias{cropland_world_2003_era5}
\title{Global cropland weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 585724 rows and 3 columns.
}
\usage{
cropland_world_2003_era5
}
\description{
A data.table with ERA5 resolution returned by running secondary_weights() with cropland data from
2003.
}
\keyword{datasets}
