% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_era5.R
\docType{data}
\name{data_era5}
\alias{data_era5}
\title{Demo ERA5 Data}
\format{
An object of class \code{RasterLayer} of dimension 721 x 1440 x 1.
}
\usage{
data_era5
}
\description{
Demo ERA5 Data
}
\keyword{datasets}
