% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_polygon.R
\docType{data}
\name{data_polygon}
\alias{data_polygon}
\title{A Demo Simple Feature Object}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 105 rows and 18 columns.
}
\usage{
data_polygon
}
\description{
A Demo Simple Feature Object
}
\keyword{datasets}
