% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prcp.R
\docType{data}
\name{data_prcp}
\alias{data_prcp}
\title{A data set of prcp data}
\format{
An object of class \code{RasterBrick} of dimension 4 x 4 x 168.
}
\usage{
data_prcp
}
\description{
A data set of prcp data
}
\keyword{datasets}
