% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weights.R
\docType{data}
\name{data_weights}
\alias{data_weights}
\title{A Demo Secondary Weights File}
\format{
An object of class \code{RasterLayer} of dimension 2000 x 2000 x 1.
}
\usage{
data_weights
}
\description{
A Demo Secondary Weights File
}
\keyword{datasets}
