% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_dual.R
\docType{data}
\name{demo_dual_weights}
\alias{demo_dual_weights}
\title{A Weight Table Calculated Using Secondary Weights}
\format{
A data table containing 42 observations of 5 variables:
\describe{
\item{x}{longitude of cell based on 0 - 360 crs}
\item{y}{latitude of cell}
\item{poly_id}{unique identifyer for each county}
\item{w_area}{proportion of the county falling within the cell}
\item{weight}{secondary weights normalized by area}
}
}
\usage{
demo_dual_weights
}
\description{
A data table containing results from calc_geoweights() when a secondary
weight table is included.
}
\keyword{datasets}
