% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5_data.R
\docType{data}
\name{demo_era5}
\alias{demo_era5}
\title{Small sample of ERA5 Data}
\format{
A raster object containing data for one year, half a degree of
longitude, and 180.5 degrees of latitude.
}
\usage{
demo_era5
}
\description{
One band of an ERA5 raster to set the default resolution and default
coordinate reference system.
}
\keyword{datasets}
