% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_sf.R
\docType{data}
\name{demo_polygon}
\alias{demo_polygon}
\title{A Simple Feature Object}
\format{
A data frame containing 105 observations of 18 variables.
}
\usage{
demo_polygon
}
\description{
Polygon data for county boundaries in Kansas, US, to be used in example of calc_geoweights().
}
\keyword{datasets}
