% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data.R
\docType{data}
\name{demo_weights}
\alias{demo_weights}
\title{Small sample of crop data}
\format{
A raster object containing data for one year, for an area that is one
degree of longitude and 1 degree of latitude across.
}
\usage{
demo_weights
}
\description{
A raster containing a small sample of crop data to be used as a secondary
weight in the calc_raster_weights() example.
}
\keyword{datasets}
