% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_weights_table.R
\docType{data}
\name{demo_weights_table}
\alias{demo_weights_table}
\title{A Secondary Weights Table Produced by Running calc_raster_weights()}
\format{
A data table containing 16 observations of 3 variables:
\describe{
\item{x}{longitude}
\item{y}{latitude}
\item{weight}{percent crop cover in a cell}
}
}
\usage{
demo_weights_table
}
\description{
A data table containing crop weights for an area 1 degree of longitude by 1
degree of latitude across to be used in an example for calc_geoweights()
}
\keyword{datasets}
