% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_era5_grid.R
\docType{data}
\name{era5_grid}
\alias{era5_grid}
\title{Empty ERA5 grid.}
\format{
An object of class \code{RasterLayer} of dimension 721 x 1440 x 1.
}
\usage{
era5_grid
}
\description{
An empty raster layer of global extent with the spatial resolution of ERA5 data to be used in
resampling a secondary raster.
}
\keyword{datasets}
