% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kansas_counties.R
\docType{data}
\name{kansas_counties}
\alias{kansas_counties}
\title{Counties of Kansas}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 105 rows and 18 columns.
}
\usage{
kansas_counties
}
\description{
A simple features object from the tigris package depicting counties in Kansas
}
\keyword{datasets}
