% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nj_counties.R
\docType{data}
\name{nj_counties}
\alias{nj_counties}
\title{Polygons representing counties in the state of New Jersey.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 21 rows and 18 columns.
}
\usage{
nj_counties
}
\description{
A multipolygon object depicting county boundaries in New Jersey, using census
data available through the tigris package.
}
\keyword{datasets}
