% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_geoweights_area.R
\docType{data}
\name{output_geoweights_area}
\alias{output_geoweights_area}
\title{An Output Table Using calc_geoweights() Without Secondary Weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 5 columns.
}
\usage{
output_geoweights_area
}
\description{
An Output Table Using calc_geoweights() Without Secondary Weights
}
\keyword{datasets}
