% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_geoweights_both.R
\docType{data}
\name{output_geoweights_both}
\alias{output_geoweights_both}
\title{An Output Table Using calc_geoweights() with Secondary Weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
output_geoweights_both
}
\description{
An Output Table Using calc_geoweights() with Secondary Weights
}
\keyword{datasets}
