% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_weights.R
\docType{data}
\name{output_weights}
\alias{output_weights}
\title{A Sample Weights Table from calc_raster_weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
output_weights
}
\description{
A Sample Weights Table from calc_raster_weights
}
\keyword{datasets}
