% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Function to find spatial overlap between a raster and a set of polygons}
\usage{
overlay(polygons, polygon_id, grid, secondary_weights = NULL)
}
\arguments{
\item{polygons}{a simple features polygon or multipolygon object}

\item{polygon_id}{the name of a column in the sf object representing a unique
identifier for each polygon}

\item{grid}{a raster layer with the same spatial resolution as the data}

\item{secondary_weights}{an optional data table of secondary weights, output
from secondary_weights()}
}
\value{
a data.table of area weights and possibly secondary weights for each
cell within polygons (area weighted raster/polygon overlap)
}
\description{
Function to find spatial overlap between a raster and a set of polygons
}
\examples{
overlay(data_era5, data_polygon, "countyfp", output_weights)
overlay(data_era5, data_polygon, "countyfp")

}
