% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overlay_weights_kansas.R
\docType{data}
\name{overlay_weights_kansas}
\alias{overlay_weights_kansas}
\title{Area and cropland weights in ERA5 grid for Kansas counties}
\format{
a data.table with 858 observations of 5 variables
\describe{
\item{x}{longitude, ranging from 0 to 360}
\item{y}{latitude, ranging from -90 to 90}
\item{polygon_id}{unique identifiers for counties in Kansas}
\item{w_area}{the proportion of the polygon that falls within a cell}
\item{weight}{value to weight by in aggregation, detirmined by
multiplying the w_area value of a cell by its secondary
weight}
}
}
\usage{
overlay_weights_kansas
}
\description{
A table of weights returned by running overlay_weights() with kansas_counties and
secondary cropland weights.
}
\keyword{datasets}
