% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_doc.R
\docType{data}
\name{poly}
\alias{poly}
\title{A Polygon}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 105 rows and 18 columns.
}
\usage{
poly
}
\description{
A Polygon
}
\keyword{datasets}
