% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cropland_world_2003_era5.R,
%   R/data_pop_world_2015_era5.R
\docType{data}
\name{pop_world_2015_era5}
\alias{pop_world_2015_era5}
\title{Global cropland weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1036800 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1036800 rows and 3 columns.
}
\usage{
pop_world_2015_era5

pop_world_2015_era5
}
\description{
A data.table with ERA5 resolution returned by running secondary_weights() with cropland data from
2003.

A data.table returned with ERA5 resolution by running secondary_weights()
with population data from 2015 LandScan Global Population Data
}
\keyword{datasets}
