% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prcp_kansas_dec2011_era5.R
\docType{data}
\name{prcp_kansas_dec2011_era5}
\alias{prcp_kansas_dec2011_era5}
\title{Precipitation data for Kansas in December}
\format{
A raster brick containing 720 layers, each representing one hour of
precipitation data for a total of 30 days worth of data
}
\usage{
prcp_kansas_dec2011_era5
}
\description{
A raster brick containing precipitation data from Dec 2, 2011 to Dec 31, 2011
for the entire state of Kansas
}
\keyword{datasets}
