% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_weights.R
\name{secondary_weights}
\alias{secondary_weights}
\title{Resample a raster of secondary weights}
\usage{
secondary_weights(secondary_raster, grid = era5_grid, extent = "full")
}
\arguments{
\item{secondary_raster}{a raster of a secondary variable, for example
cropland coverage or population}

\item{grid}{a raster layer with the same spatial resolution as the climate
data (must use geographic coordinates)}

\item{extent}{an optional extent to crop the secondary_raster to for faster
processing. Format must be compatible with raster::crop(). The default is "full", which
resamples the whole secondary raster without cropping.}
}
\value{
a data.table of secondary weights
}
\description{
The \code{secondary_weights()} function resamples a raster of interest to the
desired resolution and outputs a table of weights for use in the function
\code{overlay_weights()}.
}
\examples{
secondary_weights_output <- secondary_weights(
  secondary_raster = cropland_nj_2015, # A raster of cropland to resample
                                       # and convert to a data.table
  grid = era5_grid, # The grid to resample the secondary_raster to (`era5_grid`
                    # is the default)
  extent = "full" # The default, which resamples the whole secondary raster
                  # without cropping (`'full'` is the default)
  )

head(secondary_weights_output)

}
