% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagg-package.R
\docType{package}
\name{stagg-package}
\alias{stagg}
\alias{stagg-package}
\title{stagg: Spatiotemporal Aggregation for Climate Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
