% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_temp_kansas_jan_2011_era5.R
\docType{data}
\name{temp_kansas_jan_2020_era5}
\alias{temp_kansas_jan_2020_era5}
\title{Temperature data for Kansas in December}
\format{
A raster brick containing 744 layers, each representing one hour of
temperature data in degrees Celsius for a total of 31 days worth of data
}
\usage{
temp_kansas_jan_2020_era5
}
\description{
A raster brick containing temperature data from January 1, 2020 to January 31, 2020
for the entire state of Kansas
}
\keyword{datasets}
