% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_temp_nj_jun_2024.R
\docType{data}
\name{temp_nj_jun_2024_era5}
\alias{temp_nj_jun_2024_era5}
\title{Temperature data for New Jersey in June}
\format{
A raster brick containing 720 layers, each representing one hour of
temperature data in degrees Kelvin for a total of 30 days worth of data
}
\usage{
temp_nj_jun_2024_era5
}
\description{
A raster brick containing temperature data from June 1, 2024 to June 30, 2024
for the entire state of New Jersey
}
\keyword{datasets}
