% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{as.data.table.raster}
\alias{as.data.table.raster}
\title{Aggregate Climate Data}
\usage{
as.data.table.raster(
  x,
  row.names = NULL,
  optional = FALSE,
  xy = FALSE,
  inmem = raster::canProcessInMemory(x, 2),
  ...
)
}
\arguments{
\item{year}{The year to aggregate}

\item{data_raster}{The raster containing climate data}

\item{climate_var}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{trans}{A statistical transformation ('polynomial' currently supported)}

\item{trans_specs}{Specifications pertaining to selected transformation. For
'polynomial', this is an integer indicating the degree.}

\item{geoweights_table}{A table of weights which can be generated using
the function calc_geoweights()}

\item{second_weights}{A boolean indicating whether secondary weights should be used}
}
\description{
Aggregate Climate Data
}
\examples{
no_cores <- parallel::detectCores() - 1 # Calculate the number of cores. Leave one in case something else needs to be done on the same computer at the same time.
cl <- parallel::makeCluster(no_cores, type="FORK") # Initiate cluster. "FORK" means bring everything in your current environment with you.
sum_by_poly_multiyear <- parallel::parLapply(cl, years, agg_climate_data, data_raster, climate_var, daily_agg, trans, trans_specs, geoweights_table)
parallel::stopCluster(cl)


}
