% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{daily_quants}
\alias{daily_quants}
\title{Calculate Quantiles in Daily Climate Data}
\usage{
daily_quants(data, overlay_weights, daily_agg, probs)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using the
function calc_geoweights()}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{probs}{A vector of probabilities (0 to 1) to be calculated as quantiles from the
data}
}
\description{
Calculate Quantiles in Daily Climate Data
}
