% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Function to find spatial overlap between a raster and a set of polygons}
\usage{
overlay(polygons, polygon_id, grid, secondary_weights = NULL)
}
\arguments{
\item{polygons}{a simple features polygon or multipolygon object}

\item{polygon_id}{the name of a column in the sf object representing a unique
identifier for each polygon}

\item{grid}{the source of climate data (default is era5)}

\item{secondary_weights}{an optional data table of secondary weights}
}
\value{
a data.table of geoweights (area weighted raster/polygon overlap)
}
\description{
Function to find spatial overlap between a raster and a set of polygons
}
\examples{
overlay(data_era5, data_polygon, "countyfp", output_weights)
overlay(data_era5, data_polygon, "countyfp")

}
