% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_weights.R
\name{overlay_weights}
\alias{overlay_weights}
\title{Find the spatial overlap between a raster and a set of polygons}
\usage{
overlay_weights(
  polygons,
  polygon_id_col,
  grid = era5_grid,
  secondary_weights = NULL
)
}
\arguments{
\item{polygons}{a simple features polygon or multipolygon object}

\item{polygon_id_col}{the name of a column in the \code{polygons} object with a
unique identifier for each polygon}

\item{grid}{a raster layer with the same spatial resolution as the data
(must use geographic coordinates)}

\item{secondary_weights}{an optional table of secondary weights, output from
the \code{secondary_weights()} function}
}
\value{
a data.table of area weights and possibly secondary weights for each
cell within each polygon
}
\description{
The \code{overlay_weights()} function generates a table of weights mapping
each grid cell to its respective polygon(s) for use in the \verb{staggregate_*()}
family of functions.
}
\examples{
overlay_output_with_secondary_weights <- overlay_weights(
  polygons = tigris::counties("nj"), # Polygons outlining the 21 counties of New Jersey
  polygon_id_col = "COUNTYFP", # The name of the column with the unique
                               # county identifiers
  grid = era5_grid, # The grid to use when extracting area weights (era5_grid is the
                    # default)
  secondary_weights = cropland_world_2015_era5 # Output from
                                               # secondary_weights
                                               # (cropland_world_2015_era5 is
                                               # available to the# user)
  )

head(overlay_output_with_secondary_weights)



overlay_output_without_secondary_weights <- overlay_weights(
  polygons = tigris::counties("nj"), # Polygons outlining the 21 counties of New Jersey
  polygon_id_col = "COUNTYFP" # The name of the column with the unique county
                              # identifiers
  )

head(overlay_output_without_secondary_weights)


}
