% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_weights.R
\name{overlay_weights}
\alias{overlay_weights}
\title{Function to find spatial overlap between a raster and a set of polygons}
\usage{
overlay_weights(
  polygons,
  polygon_id_col,
  grid = era5_grid,
  secondary_weights = NULL
)
}
\arguments{
\item{polygons}{a simple features polygon or multipolygon object}

\item{polygon_id_col}{the name of a column in the sf object with a unique
identifier for each polygon}

\item{grid}{a raster layer with the same spatial resolution as the data}

\item{secondary_weights}{an optional data.table of secondary weights, output
from secondary_weights()}
}
\value{
a data.table of area weights and possibly secondary weights for each
cell within polygons (area weighted raster/polygon overlap)
}
\description{
Function to find spatial overlap between a raster and a set of polygons
}
\examples{
kansas_counties <- tigris::counties("Kansas")

overlay_output_with_secondary_weights <- overlay_weights(
  kansas_counties, # Polygons outlining the 105 counties of Kansas
  "COUNTYFP", # The name of the column with the unique county identifiers
  era5_grid, # The empty grid to resample to and align with
  cropland_world_2003_era5 # Output from secondary_weights
  )

head(overlay_output_with_secondary_weights)



overlay_output_without_secondary_weights <- overlay_weights(
  kansas_counties, # Polygons outlining the 105 counties of Kansas
  "COUNTYFP" # The name of the column with the unique county identifiers
  )

head(overlay_output_without_secondary_weights)


}
