% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pop_world_2015_era5.R
\docType{data}
\name{pop_world_2015_era5}
\alias{pop_world_2015_era5}
\title{Global population weights}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 380458 rows and 3 columns.
}
\source{
\url{https://landscan.ornl.gov/}
}
\usage{
pop_world_2015_era5
}
\description{
A data.table returned with ERA5 resolution by running secondary_weights()
with population data from 2015 LandScan Global Population Data
}
\keyword{datasets}
