% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_weights.R
\name{secondary_weights}
\alias{secondary_weights}
\title{Function that resamples a raster of interest to the ERA grid}
\usage{
secondary_weights(secondary_raster, grid, extent = "full")
}
\arguments{
\item{secondary_raster}{a raster of a continuous variable, for example cropland coverage or population}

\item{grid}{the source of climate data (default is era5)}

\item{extent}{an optional extent to crop the secondary_raster to for faster processing}
}
\value{
a data.table of secondary weights
}
\description{
Function that resamples a raster of interest to the ERA grid
}
\examples{
secondary_weights(data_weights, data_era5)

}
