% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_bin}
\alias{staggregate_bin}
\title{Bin Transformation and Aggregation of Climate Data}
\usage{
staggregate_bin(
  data,
  overlay_weights,
  daily_agg,
  time_agg = "month",
  bin_breaks
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using the
function overlay_weights()}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported) 'polynomial', this is an integer indicating the degree.}

\item{time_agg}{The temporal scale to aggregate data to ('day', 'month', and
'year' currently supported)}

\item{bin_breaks}{A vector of bin boundaries to split the data by}
}
\description{
Bin Transformation and Aggregation of Climate Data
}
\examples{
bin_output <- staggregate_bin(

  data = temp_kansas_jan_2020_era5, # Climate data to transform and aggregate

  overlay_weights = overlay_weights_kansas, # Output from overlay_weights()

  daily_agg = "average", # Average hourly values to produce daily values before transformation

  bin_breaks = c(0, 2.5, 5, 7.5, 10) # Draw 6 bins from ninf to 0, 0 to 2.5,
                                     # 2.5 to 5, 5 to 7.5, 7.5 to 10, 10 to inf
  )


head(bin_output)

}
