% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_bin}
\alias{staggregate_bin}
\title{Bin Transformation and Aggregation of Climate Data}
\usage{
staggregate_bin(
  data,
  overlay_weights,
  variable,
  daily_agg,
  agg_to = "month",
  num_bins = 10,
  binwidth = NULL,
  min = NULL,
  max = NULL,
  start_on = NULL,
  center_on = NULL,
  end_on = NULL
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using
the function calc_geoweights()}

\item{variable}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)
'polynomial', this is an integer indicating the degree.}

\item{num_bins}{number of bins to group the data into}

\item{binwidth}{width of bins, overrides num_bins}

\item{min}{the smallest value that must be captured by a non-edge bin, default is the data minimum}

\item{max}{the largest value that must be captured by a non-edge bin, default is the data maximum}

\item{start_on}{where to place the left edge of one of the bins}

\item{center_on}{where to place the center of one of the bins}

\item{end_on}{where to place the right edge of one of the bins}
}
\description{
Bin Transformation and Aggregation of Climate Data
}
\examples{
staggregate_bin(prcp_kansas_dec2011_era5, overlay_weights_kansas, variable = "prcp", daily_agg = "sum", binwidth = 2, min = 0, max = 10)

}
