% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_degree_days}
\alias{staggregate_degree_days}
\title{Degree Days Transformation and Aggregation of Climate Data}
\usage{
staggregate_degree_days(data, overlay_weights, time_agg = "day", thresholds)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using the
function overlay_weights()}

\item{time_agg}{The temporal scale to aggregate data to ('hour', 'day', 'month', and
'year' currently supported)}

\item{thresholds}{A vector of temperature thresholds critical to a crop}
}
\description{
Degree Days Transformation and Aggregation of Climate Data
}
\examples{
degree_days_output <- staggregate_degree_days(

  data = temp_kansas_jan_2020_era5, # Climate data to transform and aggregate

  overlay_weights = overlay_weights_kansas, # Output from overlay_weights()

  thresholds = c(0, 10, 20) # Calculate degree days above these 0, 10, and 20 degrees C
  )


head(degree_days_output)

}
