% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_polynomial}
\alias{staggregate_polynomial}
\title{Polynomial Transformation and Aggregation of Climate Data}
\usage{
staggregate_polynomial(
  data,
  overlay_weights,
  daily_agg,
  time_agg = "month",
  degree
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using
the function calc_geoweights()}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{time_agg}{the temporal scale to aggregate data to ('day', 'month', and
'year' currently supported)}

\item{degree}{the highest exponent to raise the data to}
}
\description{
Polynomial Transformation and Aggregation of Climate Data
}
\examples{
polynomial_output <- staggregate_polynomial(

  data = prcp_kansas_dec2011_era5, # Climate data to transform and aggregate

  overlay_weights = overlay_weights_kansas, # Output from overlay_weights()

  daily_agg = "sum", # Sum hourly values to produce daily values before transformation

  degree = 4 # Highest order
  )


head(polynomial_output)

}
