% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_polynomial}
\alias{staggregate_polynomial}
\title{Polynomial Transformation and Aggregation of Climate Data}
\usage{
staggregate_polynomial(
  data,
  overlay_weights,
  variable,
  daily_agg,
  agg_to = "month",
  degree
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using
the function calc_geoweights()}

\item{variable}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{degree}{the highest exponent to raise the data to}
}
\description{
Polynomial Transformation and Aggregation of Climate Data
}
\examples{
staggregate_polynomial(prcp_kansas_dec2011_era5, overlay_weights_kansas, variable = "prcp", daily_agg = "sum", degree = 3)

}
