% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_polynomial}
\alias{staggregate_polynomial}
\title{Polynomial transformation and aggregation of climate data}
\usage{
staggregate_polynomial(
  data,
  overlay_weights,
  daily_agg,
  time_agg = "month",
  degree
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using the
function \code{overlay_weights()}}

\item{daily_agg}{How to aggregate hourly values to daily values prior to
transformation. Options are \code{'sum'}, \code{'average'}, or \code{'none'} (\code{'none'}
will transform values without first aggregating to the daily level)}

\item{time_agg}{the temporal scale to aggregate data to. Options are
\code{'hour'}, \code{'day'}, \code{'month'}, or \code{'year'} (\code{'hour'} cannot be selected
unless \code{daily_agg = 'none'})}

\item{degree}{the highest exponent to raise the data to}
}
\description{
The function \code{staggregate_polynomial()} aggregates climate data to the daily
level, raises these daily values to the 1 through nth power, and aggregates
the transformed values to the polygon level and desired temporal scale.
}
\examples{
polynomial_output <- staggregate_polynomial(
  data = temp_nj_jun_2024_era5 - 273.15, # Climate data to transform and aggregate
  overlay_weights = overlay_weights_nj, # Output from overlay_weights()
  daily_agg = "average", # Average hourly values to produce daily values
                         # before transformation
  time_agg = "month", # Sum the transformed daily values across months
  degree = 4 # Highest order
  )

head(polynomial_output)

}
