% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_spline}
\alias{staggregate_spline}
\title{Restricted Cubic Spline Transformation and Aggregation of Climate Data}
\usage{
staggregate_spline(
  data,
  overlay_weights,
  variable,
  daily_agg,
  time_agg = "month",
  knot_locs
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using
the function calc_geoweights()}

\item{variable}{The  climate variable of interest ('prcp, 'temp', and 'uv'
currently supported)}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{time_agg}{the temporal scale to aggregate data to ('day', 'month', and
'year' currently supported)
'polynomial', this is an integer indicating the degree.}

\item{degree}{the highest exponent to raise the data to}
}
\description{
Restricted Cubic Spline Transformation and Aggregation of Climate Data
}
\examples{
staggregate_spline(prcp_kansas_dec2011_era5, overlay_weights_kansas,
variable = "prcp", daily_agg = "sum", knot_locs = c(1, 2, 3, 4, 5))

}
