% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggregate.R
\name{staggregate_spline}
\alias{staggregate_spline}
\title{Restricted Cubic Spline Transformation and Aggregation of Climate Data}
\usage{
staggregate_spline(
  data,
  overlay_weights,
  daily_agg,
  time_agg = "month",
  knot_locs
)
}
\arguments{
\item{data}{The raster brick with the data to be transformed and aggregated}

\item{overlay_weights}{A table of weights which can be generated using
the function calc_geoweights()}

\item{daily_agg}{How to aggregate daily values ('sum' and 'average' currently
supported)}

\item{time_agg}{the temporal scale to aggregate data to ('day', 'month', and
'year' currently supported)
'polynomial', this is an integer indicating the degree.}

\item{knot_locs}{where to place the knots}
}
\description{
Restricted Cubic Spline Transformation and Aggregation of Climate Data
}
\examples{
spline_output <- staggregate_spline(

data = prcp_kansas_dec2011_era5, # Climate data to transform and aggregate

overlay_weights = overlay_weights_kansas, # Output from overlay_weights()

daily_agg = "sum", # Sum hourly values to produce daily values before transformation

knot_locs = c(-1.665335e-16,  1.129775e-06,  1.594356e-02) # Where to place knots
)


head(spline_output)

}
